refactor: lift internal/{logging,observability} to public packages
Downstream binaries that assemble their own cyoda-go app instance (e.g.
cyoda-go-cassandra's full binary) need to call logging.Init(level) and
observability.Init(ctx, serviceName, nodeID) with exactly the same
semantics cyoda-go's own main.go does. Leaving these under internal/
blocks that use case without offering any real encapsulation benefit —
both packages are small, stable, and meant to be the standard setup
path for any cyoda-go-based binary.

Moved as-is; all in-tree callers had their imports rewritten. No API
changes.
