+++ internal/api/admin_test.go
@@ -10,8 +10,8 @@ import (
 
 	spi "github.com/cyoda-platform/cyoda-go-spi"
 	"github.com/cyoda-platform/cyoda-go/internal/api"
-	"github.com/cyoda-platform/cyoda-go/internal/logging"
-	"github.com/cyoda-platform/cyoda-go/internal/observability"
+	"github.com/cyoda-platform/cyoda-go/logging"
+	"github.com/cyoda-platform/cyoda-go/observability"
 )
 
 // adminContext returns a request with a ROLE_ADMIN UserContext attached.
